requirements-dev.txt
# This file is typically generated by 'pip-compile requirements-dev.in'.
# It lists all direct and indirect dependencies with pinned versions for reproducible development environments.
# Example versions are provided; actual versions should be generated.

# --- Core ML / CV ---
# Assuming TensorFlow 2.x. For specific versions, check compatibility with CUDA/cuDNN if using GPU.
tensorflow==2.16.1 # Or tensorflow-cpu==2.16.1
numpy==1.26.4
scipy==1.13.0
opencv-python==4.9.0.80
pandas==2.2.2
matplotlib==3.8.4
Pillow==10.3.0

# --- Development Utilities ---
jupyterlab==4.2.1
  # jupyterlab dependencies (will be many, e.g., jupyter-server, tornado, etc.)
pylint==3.1.0
  # pylint dependencies
pytest==8.2.0
  # pytest dependencies
black==24.4.2
flake8==7.0.0
mypy==1.10.0

# --- Hardware Interfacing ---
pyserial==3.5

# --- Other transitive dependencies will be listed here by pip-compile ---
# Example: (These are illustrative, pip-compile will resolve actual dependencies)
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
Keras==2.15.0 # Or keras==3.3.3 if using Keras 3 with TF 2.16
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
ml-dtypes==0.2.0 # Or a newer version compatible with TF
opt-einsum==3.3.0
packaging==24.0
protobuf==4.25.3 # Or a version compatible with TF
requests==2.31.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
typing-extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
# ... and many more

# --- Optional Additions for Enhanced Development (Consider adding if needed) ---
scikit-image             # For a wider range of image processing algorithms
albumentations           # Powerful library for advanced data augmentation
tqdm                     # For progress bars during long operations (e.g., data processing, training)
python-dotenv            # For managing environment variables if using a .env file for API keys/configs
